+++ kamehouse-admin/src/main/java/com/nicobrest/kamehouse/admin/model/kamehousecommand/KameHouseDesktopStartupKameHouseCommand.java
@@ -3,17 +3,22 @@ package com.nicobrest.kamehouse.admin.model.kamehousecommand;
 import com.nicobrest.kamehouse.commons.model.kamehousecommand.KameHouseShellScript;
 
 /**
- * KameHouse command to restart kamehouse desktop.
+ * KameHouse command to start kamehouse desktop.
  *
  * @author nbrest
  */
-public class KameHouseDesktopRestartKameHouseCommand extends KameHouseShellScript {
+public class KameHouseDesktopStartupKameHouseCommand extends KameHouseShellScript {
 
   /**
    * Sets the command line for each operation required for this KameHouseCommand.
    */
-  public KameHouseDesktopRestartKameHouseCommand() {
+  public KameHouseDesktopStartupKameHouseCommand() {
     super();
+  }
+
+  @Override
+  public boolean isDaemon() {
+    return true;
   }
 
   @Override
@@ -28,11 +33,11 @@ public class KameHouseDesktopRestartKameHouseCommand extends KameHouseShellScrip
 
   @Override
   protected String getWindowsKameHouseShellScript() {
-    return "desktop/kamehouse-desktop-restart.sh";
+    return "desktop/kamehouse-desktop-startup.sh";
   }
 
   @Override
   protected String getLinuxKameHouseShellScript() {
-    return "desktop/kamehouse-desktop-restart.sh";
+    return "desktop/kamehouse-desktop-startup.sh";
   }
 }